fix(react-jss): rename file becuase aim changed
